feat: Integrate Netlify with antora (#74)

+++ src/utils/serde/u64_deserializer.rs
@@ -30,7 +30,7 @@ impl de::Visitor<'_> for U64Visitor {
     where
         E: de::Error,
     {
-        Ok(value as u64)
+        Ok(value)
     }
 
     // Handle i64 inputs
